# perf(chstorage): do not make an unnecessary copy of encoded attributes