For #293.

```          
                 │   old.txt    │               new.txt               │
                 │    sec/op    │   sec/op     vs base                │
_metricsBatch-32   1059.2µ ± 0%   824.0µ ± 1%  -22.21% (p=0.000 n=15)

                 │    old.txt     │               new.txt                │
                 │      B/op      │     B/op      vs base                │
_metricsBatch-32   290.716Ki ± 0%   4.317Ki ± 0%  -98.51% (p=0.000 n=15)

                 │   old.txt    │              new.txt               │
                 │  allocs/op   │ allocs/op   vs base                │
_metricsBatch-32   6280.00 ± 0%   16.00 ± 0%  -99.75% (p=0.000 n=15)
```